Adjustments to adhere to strict ISO standards.

+++ module-interface/include/intf_io.h
@@ -6,7 +6,7 @@
 
 #define DEFAULT_BUF_SIZE 32
 
-void intf_io_fgets(char *buffer, int size);
+void intf_io_fgets(char *buffer, size_t size);
 int intf_io_opt();
 
 #endif //REPAIRSHOP_INTF_IO_H
